fix: change WeatherResponse.wind_speed from int to float
OpenWeatherMap returns fractional wind speeds (e.g. 2.06 m/s).
Pydantic v2 strict int validation rejected these, causing a 500 error
on GET /api/v1/dashboard/weather.

+++ system_a/app/api/v1/dashboard_widgets.py
@@ -281,7 +281,7 @@ class WeatherResponse(BaseModel):
     temperature: float = 0  # °C from ambient sensor
     condition: str = "sunny"  # sunny, cloudy, rainy, windy
     humidity: int = 50
-    wind_speed: int = 10  # km/h
+    wind_speed: float = 10.0  # m/s
     solar_forecast: int = 0  # % of expected production
     sunrise: str = "06:00"
     sunset: str = "18:00"
